+++ apps/backend/package.json
@@ -6,9 +6,10 @@
   },
   "dependencies": {
     "drizzle-orm": "^0.44.2",
-    "hono": "^4.7.11"
+    "hono": "^4.8.3"
   },
   "devDependencies": {
-    "@types/bun": "latest"
+    "@cloudflare/workers-types": "^4.20250628.0",
+    "@types/bun": "^1.2.17"
   }
 }
